move getValidRowHeight to gridRowsUtils
